+++ bundles/basic/webpack.config.js
@@ -1,0 +1,4 @@
+const { loadParticlesBundle } = require("@tsparticles/webpack-plugin");
+const version = require("./package.json").version;
+
+module.exports = loadParticlesBundle({ moduleName: "basic", bundleName: "Basic", version, dir: __dirname });