+++ engine/src/Core/Canvas.ts
@@ -55,7 +55,7 @@ export class Canvas {
 
     private _coverColorStyle?: string;
     private _generated;
-    private readonly _mutationObserver?: MutationObserver;
+    private _mutationObserver?: MutationObserver;
     private _originalStyle?: CSSStyleDeclaration;
     private _postDrawUpdaters: IParticleUpdater[];
     private _preDrawUpdaters: IParticleUpdater[];
@@ -78,13 +78,6 @@ export class Canvas {
         this._postDrawUpdaters = [];
         this._resizePlugins = [];
         this._colorPlugins = [];
-        this._mutationObserver = safeMutationObserver((records) => {
-            for (const record of records) {
-                if (record.type === "attributes" && record.attributeName === "style") {
-                    this._repairStyle();
-                }
-            }
-        });
     }
 
     private get _fullScreen(): boolean {
@@ -118,7 +111,7 @@ export class Canvas {
      * Destroying object actions
      */
     destroy(): void {
-        this._safeMutationObserver((obs) => obs.disconnect());
+        this.stop();
 
         if (this._generated) {
             const element = this.element;
@@ -127,8 +120,6 @@ export class Canvas {
         } else {
             this._resetOriginalStyle();
         }
-
-        this.stop();
 
         this._preDrawUpdaters = [];
         this._postDrawUpdaters = [];
@@ -243,6 +234,14 @@ export class Canvas {
      * Initializes the canvas element
      */
     async init(): Promise<void> {
+        this._safeMutationObserver((obs) => obs.disconnect());
+        this._mutationObserver = safeMutationObserver((records) => {
+            for (const record of records) {
+                if (record.type === "attributes" && record.attributeName === "style") {
+                    this._repairStyle();
+                }
+            }
+        });
         this.resize();
         this._initStyle();
         this._initCover();
@@ -426,6 +425,9 @@ export class Canvas {
     }
 
     stop(): void {
+        this._safeMutationObserver((obs) => obs.disconnect());
+        this._mutationObserver = undefined;
+
         this.draw((ctx) => clear(ctx, this.size));
     }