+++ updaters/outModes/src/DestroyOutMode.ts
@@ -22,7 +22,7 @@ export class DestroyOutMode implements IOutModeManager {
     update(
         particle: Particle,
         direction: OutModeDirection,
-        delta: IDelta,
+        _delta: IDelta,
         outMode: OutMode | OutModeAlt | keyof typeof OutMode,
     ): void {
         if (!this.modes.includes(outMode)) {